+++ acceptance-tests/src/test/java/tech/pegasys/web3signer/dsl/signer/runner/CmdLineParamsDefaultImpl.java
@@ -321,6 +321,11 @@ public class CmdLineParamsDefaultImpl implements CmdLineParamsBuilder {
       params.add(Long.toString(signerConfig.getFuluForkEpoch().get()));
     }
 
+    if (signerConfig.getGloasForkEpoch().isPresent()) {
+      params.add("--Xnetwork-gloas-fork-epoch");
+      params.add(Long.toString(signerConfig.getGloasForkEpoch().get()));
+    }
+
     if (signerConfig.getNetwork().isPresent()) {
       params.add("--network");
       params.add(signerConfig.getNetwork().get());